Add bash hook
