+++ airflow/example_dags/example_bash_operator.py
@@ -21,7 +21,8 @@
 from datetime import timedelta
 
 from airflow import DAG
-from airflow.operators.bash import EXIT_CODE_SKIP, BashOperator
+from airflow.operators.bash import BashOperator
+from airflow.hooks.bash import EXIT_CODE_SKIP
 from airflow.operators.dummy import DummyOperator
 from airflow.utils.dates import days_ago
